Introduce examples-common module to host code shared between java and kotlin agents
