+++ embabel-agent-examples/examples-common/src/main/kotlin/com/embabel/examples/simple/horoscope/HoroscopeService.kt
